[autotest] Change update command execution in deploy_production_local.
Various commands in the [UPDATE_COMMANDS] section in
global_config.ini need to refer to content in the local autotest
repository.  Previously, that was done by having
deploy_server_local replace the special string "AUTOTEST_REPO" with
the path to the directory.

This change drops the old behavior in favor of just allowing
commands in [UPDATE_COMMANDS] to assume that the current directory
is the Autotest repository.

BUG=None
TEST=TBD

Change-Id: Ia3e24a8f99151f87ce111fb42ddb55c9361500cc
Reviewed-on: https://chromium-review.googlesource.com/230940
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>