faft: Fix modifying setvars.sh to update the RO_FRID
When running FAFT key-update tests without giving an extra shellball,
it will replace the firmware binary of the shellball with the
existing active firmware. In the unibuild design, the FWID infos are
recorded in the setvars.sh. We should modify the setvars.sh to
reflect the binary replacement.

There was a bug that the RO_FRID was updated wrong, the same value as
the RW_FWID. So if we use a different version on RW, the key-update
tests will fail.

This fix updates the RO_FRID to a correct value, by reading the
RO_FRID section of the active firmware.

BUG=b:72814947
TEST=Ran firmware_UpdateFirmwareDataKeyVersion on Fizz.

Change-Id: I3325c90cfa1fbf50731dfdda3fdb349ac3694eb5
Reviewed-on: https://chromium-review.googlesource.com/959240
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>