autotest: lxc: use lxc-copy where available
On shards, we have lxc-copy, which is preferred.
On moblab, we still use an old version of lxc, which does not have
lxc-copy and still uses the deprecated lxc-clone.
This checks for the existence of lxc-copy, uses it if able, and falls
back to the old command if necessary.

BUG=chromium:822112
TEST=Hwtest tryjob on moblab

Change-Id: Ie0dbe9048ef503db371a3f74c42b81ec6c6767b0
Reviewed-on: https://chromium-review.googlesource.com/965398
Commit-Ready: Jacob Kopczynski <jkop@chromium.org>
Tested-by: Jacob Kopczynski <jkop@chromium.org>
Reviewed-by: Laurence Goodby <lgoodby@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>