firmware_DevMode: Add test for dev-mode transition when EC_IN_RW is set
The current firmware_DevMode test checks that the dev mode transition
works when it's supposed to work, but it doesn't check that it won't
work when it's not supposed to work. Considering that the EC_IN_RW latch
is an important part of our security story, we should probably have a
full-system integration test to check that it works as intended. This
patch adds a step to firmware_DevMode that will attempt to transition
into developer mode with a "compromised EC" (an EC that signals the
"recovery mode" host event even though it is not in RO) and esnures that
the system will stay in normal mode in that case.

BUG=None
TEST=Ran on Elm and Scarlet.

Change-Id: Id25f36621afa165753c77462a7d2458a0baa2cdf
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/972426
Reviewed-by: Wai-Hong Tam <waihong@google.com>