autotest: Add option to pass FileStore to autoserv
This CL adds a new argument --host-info-subdir to autoserv to provide a
subdirectory within the results directory where prepopulated HostInfo
stores reside for the targeted machines.

If provided, these stores will be used for and updated by the autoserv
execution. The default (when the option is not provided) is to
initialize stores backed by the AFE.

BUG=chromium:821087
BUG=chromium:827388
TEST=Run task on skylab-drone; Run provision on moblab-vm

Change-Id: I7bc01f135553364a91f44617c6d1cfa319442322
Reviewed-on: https://chromium-review.googlesource.com/974361
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>