[autotest] liststorage: Handling '\040' (space) unicode in mountpoint
In the current scenario, a default case mounts the usb device without
logging in and execute file operation. This is not true in a true user
scenario.

After login, a non-labelled device is mounted to
'/media/removable/USB\ Drive'. When read by python, an octal unicode
'\040' replaces ' ' (space) in the above mentioned path. This patch
escapes all such octal unicodes.

BUG=None
BRANCH=None
TEST= "
Tested using 'hardware_UsbBasicFileOperations':
Without login - the case passes with no issues as it executes file
operations on the device mounted to a default mountpoint.
With Login - the case passes as the device is mounted to
'/media/removable/USB Drive' and space is handled.
"
Signed-off-by: pharlapu <puneeth.harlapur@intel.com>

Change-Id: I9dd9948e5dad32a51e4681a34404a263a6ea27de
Reviewed-on: https://chromium-review.googlesource.com/974986
Commit-Ready: Puneeth Harlapur <puneeth.harlapur@intel.com>
Tested-by: Naresh Solanki <naresh.solanki@intel.com>
Tested-by: Puneeth Harlapur <puneeth.harlapur@intel.com>
Reviewed-by: Naresh Solanki <naresh.solanki@intel.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Bogineni Kasaiah <kasaiahx.bogineni@intel.com>
Reviewed-by: Vaishali Paisal <vaishali.paisal@intel.com>