cryptohome: Fix autotest's TPM own waiting
This fixes the "wait_for_ownership" feature of the take_tpm_ownership()
helper to not return preliminarily.

This could happen because the 'Owned' flag is actually set by
cryptohomed before the ownership tasks are completed. In other words,
this status response is entirely possible today, and was causing issues
with the old implementation:
  {'Ready': False,
   'Being Owned': True,
   'Password': '',
   'Enabled': True,
   'Owned': True}

The fix is to consider the 'Being Owned' flag too. This flag is released
by cryptohomed at the very end of the ownership procedure.

BUG=chromium:806788
TEST=none

Change-Id: I92947497b1cf21684ed1af0791167fac10f6bb0e
Reviewed-on: https://chromium-review.googlesource.com/976224
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>