[autotest] Delete a problematic check from machine_install().
In CrosHost.machine_install(), there's a sanity check that the
build installed on the DUT is the same as the build that was
requested.  The check operates by comparing strings in
/etc/lsb-release against portions of the URL that named the build;
that check requires various heuristics and special cases.

The check in question is only used by Moblab, not by Autotest in
the full lab.

A recent change to the way builds are named on the builders broke
the check in machine_install(), and now the moblab pre-CQ builder
is failing.  The check is somewhat redundant, so stop the bleeding
by deleting the check.

BUG=chromium:824956
TEST=run this through the pre-cq

Change-Id: Iccf19779b5f7a4c7a4e8e233e95c8c683cf6e990
Reviewed-on: https://chromium-review.googlesource.com/976644
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Trybot-Ready: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>