firwmare_Cr50OpenWhileAPOff: modify how devices are held in reset
Boards like bob cannot tell that the AP is off when the EC is held in
reset. During init determine the level of deep sleep support. If Cr50
cannot tell the AP is off when the EC is in reset, just rely on turning
the AP off. Skip trying to fully open cr50, because the power button
presses required to open cr50 would turn on the AP.

BUG=none
BRANCH=none
TEST=run on bob and reef. Make sure it fails with 2.1 and passes with
3.3

Change-Id: If1e24b1ee3dc5618f623114c82e5d6057f254856
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/978669
Commit-Ready: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>