[autotest] Factor out _write_keyvals()
Going to start cleaning up some of the mess after adding a significant
amount of lucifer code.

Move some stuff around so that the keyvals logic can be moved to a
separate module (this logic currently lives in the scheduler, but
lucifer is taking the job running responsibility away from the
scheduler).

BUG=chromium:810141
TEST=None

Change-Id: I4f82502d40d8e5463f8e5eff05813e0a2e55eae1
Reviewed-on: https://chromium-review.googlesource.com/978960
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Paul Hobbs <phobbs@google.com>