More tests for bad data in omaha response.
This adds on the previous autoupdate_OmahaResponse CL to add tests for
dealing with bad metadata_size and SHA256 values in the omaha response.

It is automating the test cases in section 1.14 here:
https://testtracker.googleplex.com/testplans/details/450

This CL adds a new client test autoupdate_BadMetadata that will do the
actual checking and handling of the update.

The server test (autoupdate_OmahaResponse) was updated to call this new
test.

This CL also adds a raise_error boolean to
check_update_engine_log_for_entry that I forgot to add to the last CL.

BUG=chromium:828686
TEST=autoupdate_OmahaResponse.local with all 3 tests.

Change-Id: I0e8f9fe193e07c8a5d364d7101c99f49743428cc
Reviewed-on: https://chromium-review.googlesource.com/994261
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>