Add autotests for update_engine's backoff mechanism.
Dev team added the ability to test backoff (CL:988296) and this CL tests
this functionality.

In order to test backoff we start an update, then kill the network so it
fails. Then we try to update again with backoff enabled and disabled.

In this CL:
1. Add new entry to OmahaResponse for backoff
2. Added client test autoupdate_Backoff
3. Allowed the number of urls in nano_omaha_devserver to be
configurable so we can get to the backoff testing quicker (ie we only
wait for one url failure before backoff, instead of trying both).
4. Move some code from autoupdate_UrlSwitch to base class

BUG=chromium:825065
TEST=autoupdate_ForcedOOBEUpdate, autoupdate_OmahaResponse.url_switch
TEST=autoupdate_OmahaResponse.backoff_enabled.full
TEST=autoupdate_OmahaResponse.backoff_disabled.full

Change-Id: Ib59a52008505efb2f468333795e64739a740bd1c
Reviewed-on: https://chromium-review.googlesource.com/1003769
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>