Fixing YouTubeHTML5 flakiness
YouTube uses a sophisticated algorithm to decide the playback
resolution and its Javascript API for resolution change is
merely a suggestion mechanism. As a result, we often see that
the resolution we asked for is not delivered. I have spoken with
YouTube folks and this is by design. I am removing the code we had
to test resolution changing. It still validates playback and player
states.

BUG=chromium:820136
TEST=Tested on chromeos4-row4-rack6-host1

	modified:   client/cros/video/youtube_helper.py
	modified:   client/site_tests/video_YouTubeHTML5/control
	modified:   client/site_tests/video_YouTubeHTML5/video_YouTubeHTML5.py

Change-Id: Ie0f6b3de5e08d255ddf6d23ac5416ae39df18647
Reviewed-on: https://chromium-review.googlesource.com/1006475
Commit-Ready: Vinayak Suley <vsuley@chromium.org>
Tested-by: Vinayak Suley <vsuley@chromium.org>
Reviewed-by: Krishna Gavini <krishnargv@chromium.org>
Reviewed-by: Vinayak Suley <vsuley@chromium.org>