autotest: Add graphics_VideoRenderingPower autotest
This CL adds a new autotest graphics_VideoRenderingPower,
together with an example vp9 control file; this new test
measures the average power consumption while rendering video
for a while with some different flags (hw overlay on/off).

The result is produced as an output_perf_value with the
idea of adding it to a performance dashboard (see next CL
in the chain).

Bug:chromium:831412
Change-Id: I30cfc465b474329754a948db1e833ad4c78ce17d
Reviewed-on: https://chromium-review.googlesource.com/1005082
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>