power_telemetry_logger: Fix jitter in SweetberryThread
Currently when we ask SweetberryThread to log data every
20 seconds, Measurement logger will do the following

while not done
  log power for 20 seconds
  do some log parsing

This lead to timing jitter as logging power does not finish
instantly and logging timestamp will be something like this.
0, 20.001, 40.003, 60.003, ..., 3602.345 which is multiple
seconds jitter after an hour which is not ideal data analysis
as the timestamp for each run is not matched.

This CL fixes that logic to eliminate jitter by calculate time
that eliminate the jitter.

BUG=b:69810407
TEST=No jitter in fast PLT logging every 5 seconds

Change-Id: I4a871b90bd8bf1339f6dc56e4e8722f87251e363
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1011286