Remove dependency on servo for suspend/resume in AU tests.
We currently use servo's lid_close to suspend in the AU tests. This will
not work on chromebases, as well as some chromebooks that dont support
it.

This CL changes it to use powerd which is supported by all chrome
devices.

I've also reduced the max pecentage to interrupt the update to 70%
(down from 80%) as the update was sometimes finishing before the
interruptions :)

BUG=chromium:832292
TEST=autoupdate_ForcedOOBEUpdate.interrupt.full

Change-Id: I1900712bd75bc2173ff70579258bfd348c136fb2
Reviewed-on: https://chromium-review.googlesource.com/1012988
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>