Fix for P2P AU Test failures.
The problem was that depending on when the test was ran, it would pass
or fail. If it followed a test that did a delta update it passed since
the payload signature would be correct and setting max_p2p_attempts > 10
would result in update_engine preventing any further attempts. However,
if the payload signature didn't match it would reset all state
(including p2p attempts) and the update would succeed.

This CL copies the payload signature from host 1 -> 2 so that we can
just edit the pref files to get into the state we need.

BUG=chromium:817908
TEST=auotoupdate_P2P.delta, autoupdate_P2P.deadline_expired,
autoupdate_P2P.too_many_attempts

Change-Id: I162ed699b7e82efbade3663aea30dc1e3c26fd0f
Reviewed-on: https://chromium-review.googlesource.com/1013334
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>