[autotest] Abort jobs that error when sending to Lucifer
Abort jobs that cause this error (synch_count=1 jobs with multiple
HQEs).

Its hard to handle this from Lucifer, so let the scheduler handle it
as an abort.

For example, one of the HQEs might be STARTING while the other is
PROVISIONING, and Lucifer cannot gracefully abort the PROVISIONING
HQE.

BUG=chromium:832167
TEST=Test locally (create a bad job followed by a good job to check host scheduling)

Change-Id: Iddb215d8a0669fa5fb15e507f3a87777371321b1
Reviewed-on: https://chromium-review.googlesource.com/1014395
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>