some diagnostic logging
This will check if the "'NoneType' object is not callable" in <bound
method Cursor.__del__ of <MySQLdb.cursors.Cursor object...
errors are emanating from near the INFO messages that frequently
accompany them. If so, root-causing them to be able to silence them will
be easier.

BUG=chromium:826534
TEST=None

Change-Id: Ia6f27be35238b9369cc74d11d1cf08bf5feaf38f
Reviewed-on: https://chromium-review.googlesource.com/1018529
Commit-Ready: Jacob Kopczynski <jkop@chromium.org>
Tested-by: Jacob Kopczynski <jkop@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>