autotest: Hack retrieve_logs to not redirect to shards behind uberproxy.
Some shards are not accessible via uberproxy for some users. Our check
for results on the shards was broken because it misinterpreted a
redirect to the login page as a redirect to GS. We now reject such
redirects.

The implication is that while the logs are still on the shard (before
they're offloaded to GS), users who did have access to the shard
directly will not be redirected correctly.

BUG=chromium:828902
TEST=None. Land, then have someone with/without access verify on
staging.

Change-Id: Iadbcd4506e3d818ee40a599e423ad1c2dcee865d
Reviewed-on: https://chromium-review.googlesource.com/1026561
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>