autotest: add graphics_Chrome.ozone_gl_unittests
This CL adds a new graphics_Chrome to run a given Chrome unittests
binary, and uses it to run ozone_gl_unittests. This is a tiny test
that allocates a DmaBuf and fills it in, but needs a bunch of Chrome
subsystems to be OK and is known to have regressed in the past
inadvertently.

This CL also wires the |timeout| parameter in run_chrome_test_binary()
down to its call to utils.system(), and uses it.

Bug: chromium:831811
Change-Id: I5e85c05031e8e43ce8f8478ead5fc5433a81c57e
Reviewed-on: https://chromium-review.googlesource.com/1008782
Commit-Ready: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>