video_PlaybackPerf: Add audio to VP8 4k test video
The VP8 4k video test doesn't have audio. It is different from the test videos
for H264 and VP9 4k tests.
Chrome, as of today, doesn't proceed seeking if a display is run out of and a
playing video doens't have audio in video only page (i.e. when opening a video
with chrome).
In other words, display is hitched. It shouldn't affect the number of dropped
frames, but this test would test end-to-end playback and the behavior looks
strange and misunderstandable.
This changes the test video as one with audio. Other things but audio (e.g.
bitrate and fps) is unchanged.

BUG=chromium:827912
TEST=video_PlaybackPerf.vp8.4k on eve

Change-Id: I950a79b83c3baf6477c54ef3962b9e77ba4cb7e3
Reviewed-on: https://chromium-review.googlesource.com/1027330
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Wu-Cheng Li <wuchengli@chromium.org>