graphics_Idle: Add eve to PSR tests
Previously we only had samus, but eve also has PSR, so let's add it
so it doesn't break. We also fix the test to work on newer kernels,
although the accuracy isn't as good because the same info we have
on older kernels isn't present.

BUG=none
TEST=this test on eve

Change-Id: If5c1027ca1635aa217a614259547cfd2d1d60670
Signed-off-by: Stphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1027999
Commit-Ready: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>