[ARC++] Fix the Play Store disable function
The code was using UIAutomator to navigate the Android settings app, but
this was broken by a newer version of the app which had a different
interface for navigating to the App ettings pages.

Instead we launch the Settings activity we want (Application
Info/Details) much more directly using a more standard action command
line argument, naming the package we want the settings for on the
command line.

We still assume that the UI on that page is the same, and that the same
sorts of UI actions can disable the app, but for now that does seem to
be the same in the new version.

BUG=b:78476884
TEST=test_that cheets_MailBench # passes on Eve

Change-Id: I8a8246169ead11ce1cbbea73518e6c33926433b4
Reviewed-on: https://chromium-review.googlesource.com/1033970
Commit-Ready: Lloyd Pique <lpique@chromium.org>
Tested-by: Lloyd Pique <lpique@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>