wifi_proxy: Replace mutable list argument with None
This fixes the following pylint warning:
W: 44, 4: Dangerous default value {} as argument (dangerous-default-value)

From the python docs:
(https://docs.python.org/3/tutorial/controlflow.html#default-argument-values)
The default value is evaluated only once. This makes a difference when the
default is a mutable object such as a list, dictionary, or instances of most
classes. A function with a mutable argument accumulates the arguments passed
to it on subsequent calls.

BUG=chromium:838014
TEST=None

Change-Id: I5a4bef88a911ba75b1a0eca0cc6924a45615b59a
Reviewed-on: https://chromium-review.googlesource.com/1034376
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>