[autotest] Make stable_version handle empty image type
If the stable_version command were asked to list mappings for an
image type where there were no mappings to print, the command would
fail with an exception.  This could happen in some out of the way
cases, such as with the staging instance.

This fixes the handling to print nothing when there is nothing to
print.

BUG=None
TEST=Run the command against a local instance

Change-Id: Id7e7aabd615debec366d793f0012fb25bf3d3f11
Reviewed-on: https://chromium-review.googlesource.com/1036242
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Ningning Xia <nxia@chromium.org>