servo: print gpio value as repr instead of str
This just makes it a bit easier to tell exactly what autotest is sending
to the uart. It will print the whitespace represnentations instead of
the actual white space. This condenses cmds with linebreaks, so it's
easier to see whats going on.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4c23978d4a3577d6b01d206ac3bb444817fc4e9f
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1038885
Commit-Ready: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>