Fixes for most common AU OOBE test failures.
The OOBE tests are currently failing sometimes because we are quitting after
30 seconds if the update has not started. Looking at the logs the update
does actually start during the test tear down so we just need to wait a
bit longer.

The other failure reason is a test problem that assumes a json response has
specific keys. We will check for the key.

BUG=None
TEST=None

Change-Id: I83322adc4e6b1a38dbfbe334376df9bdcd79dd57
Reviewed-on: https://chromium-review.googlesource.com/1038814
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>