autotest: Log DB Autocommit enable.
We create many database connections implicitly, and can get vague
error messages back on connection failure which don't identify WHICH
DB we failed to connect too.

This spot is often the initial connection point, and so logging the
database config name used gives a pointer towards which DB we are
attempting to connect too.

BUG=chromium:839028
TEST=None

Change-Id: Ie7780a5f5cf68051f776e792592e1fe692483f96
Reviewed-on: https://chromium-review.googlesource.com/1042891
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>