power_ProbeDriver.probe_ac: Allow battery discharge in some cases.
Some devices allow for battery discharge on AC.  In particular those
which carry the following chrome-ec CONFIG option,
CONFIG_CHARGER_DISCHARGE_ON_AC.

While we could generate a whitelist via,
  git grep CONFIG_CHARGER_DISCHARGE_ON_AC board/*/board.h
  board/reef/board.h:#define CONFIG_CHARGER_DISCHARGE_ON_AC
  ...

That would be difficult to maintain.  This CL proposes an alternative
which is to survey that:
  1. AC is in fact charging
  2. state-of-charge is> 95%

which should demonstrate that charging has in fact been successful we
just so happen to be in the discharge phase of battery maintenance.

BUG=chromium:820588
TEST=power_ProbeDriver.probe_ac passes on snappy (reef derivative)

Change-Id: Id0f3d91ea69bb0332f40d57cb42a589a40a1c54e
Reviewed-on: https://chromium-review.googlesource.com/1050701
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>