autotest: Restore Lucifer parsing code path
Originally removed by
https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/999173

I think the provision->parse code path is still going through Lucifer,
and I overlooked this, but I dont remember (I had thought I added an
exception for this case somewhere).

In any case, this code is otherwise dead, so if my hypothesis is wrong
this will be adding back dead code and do no harm.

Land this first to see if it fixes the problem and then investigate
later.

Note that this actually will cause Lucifer to crash because this code
was removed on the Lucifer side as well, but this will "unstick" jobs
by getting them marked Failed instead of getting stuck in Parsing.

BUG=chromium:841597
TEST=None

Change-Id: I759c5f1541c8f5167a5f561561c298ed207d1227
Reviewed-on: https://chromium-review.googlesource.com/1053374
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Ningning Xia <nxia@chromium.org>