autotest: Enterprise WiFi Autoconnect Tests
This test consists of two components, a server component, that creates a
wireless access point, and a client test, that sets the network
configuration policy. The client test then disconnects ethernet and verifies
that the DUT autoconnects (or doesn't) to WiFi based on the given policy.
Ethernet is re-enabled after the test is completed.

go/enterprise-networking-autotestplan

BUG=None
TEST=ran this test on lars and cyan boards in pool:wificells
CQ-DEPEND=CL:1060463

Change-Id: I02a5a51ee5f1207084603a86eb283f2c04712ced
Reviewed-on: https://chromium-review.googlesource.com/1054781
Commit-Ready: Max Timkovich <timkovich@chromium.org>
Tested-by: Max Timkovich <timkovich@chromium.org>
Reviewed-by: Max Timkovich <timkovich@chromium.org>