security_SandboxedServices: Revamp how errors and warnings are presented
This change makes it easier to update the baselines since there is now a
nice report that contains all the information, instead of having to hunt
down a series of logs. Additionally, warnings are printed even when
errors occur, so you can update all the expectations in one go instead
of doing it incrementally.

BUG=None
TEST=security_SandboxedServices

Change-Id: Id61fe340c197ebd44ad0f2f48232c5107e35adf6
Reviewed-on: https://chromium-review.googlesource.com/1060339
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>