power_suspend: Fix suspend abort check
This fixes the suspend abort check to look at a log message that
powerd generates instead of the kernel. This is because the kernel log
message is not longer representative of the error we'd like to detect.
The powerd message is based on errno and stderr, so that should be
more stable than using kernel logs.

BUG=b:79734640
TEST=power_Resume on octopus

Change-Id: I360a8a39303f8591f1df30c48d5ee89f6d2f1bc8
Reviewed-on: https://chromium-review.googlesource.com/1063319
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>