[autotest] Reland: Change to make lxc-info call work with lxc 3.0.0
There is a change to the config syntax between lxc 2.x.x and
lxc 3.x.x, determine the version of lxc we are running and
make the correct call.

I did it this way rather than just catch the error as I feel
that version differences are going to continue to happen and this
should make it easier to make future changes.

Reland fix:
Changed to ensure that the code will not crash in the event of being
called with on a machine where lxc is not installed.

BUG=chromium:844050
TEST=tested on moblab running lxc.3.x and tryjobs

Change-Id: I9ccb41ae40bbe173d5c03494aee102443ce8533e
Reviewed-on: https://chromium-review.googlesource.com/1110556
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>