[autotest] Change to make lxc-info call work with lxc 3.0.0
There is a change to the config syntax between lxc 2.x.x and
lxc 3.x.x, determine the version of lxc we are running and
make the correct call.

I did it this way rather than just catch the error as I feel
that version differences are going to continue to happen and this
should make it easier to make future changes.

BUG=chromium:844050
TEST=tested on moblab running lxc.3.x and tryjobs

Change-Id: Ib605a78858c13e464a92dcbbb63668a3b5307f54
Reviewed-on: https://chromium-review.googlesource.com/1064690
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Jacob Kopczynski <jkop@chromium.org>