[autotest] Simplify update kernel priority check.
When installing a new build, there's a sanity check applied after
downloading the build and before rebooting.  The check makes sure
that the newly installed kernel will be the kernel at next boot.

This change simplifies the logic, with the intent of making the
test easier to understand.

BUG=None
TEST=run the sanity suite in a local instance

Change-Id: Ibc51ae20de7ffb1969fce9185f031ff78ddcea09
Reviewed-on: https://chromium-review.googlesource.com/1064797
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>