autotest: wifi: Add checks to support pure VHT80 testing
As of this change, autotest doesn't have tests with MODE_11AC_PURE. We'd
like to start testing 802.11ac VHT80 rates exclusively. This change
introduces a check to enforce that the client can support these rates.
Pieces involved are:
a. Change the require_vht variable in hostap_config to make it appear
not-private.
b. Check for require_vht in the ap config, in wifi_test_context_manager,
then error out if the client does not support AC rates.
c. Remove duplicate checks from existing wifi tests, since we added one
in  wifi_test_context_manager.
network_WiFi_BluetoothStreamPerf has this check, but it currently only
runs on 802.11a and 802.11b configs. Remove it.
network_WiFi_Perf has this check, but its not the same. It enforces that the
client support 802.11ac VHT only rates, even when the AP is asking for a
"mixed" that is both 802.11ac VHT and 802.11n HT. Add a comment
explaining that, and leave the check intact.
d. Check the require_capabilities function logging in site_linux_system
to not assume that its running on an AP.

BUG=chromium:827829
TEST=Run a modified network_WiFi_SimpleConnect with VHT80 config on a client
that does not support 802.11ac, and confirm that the test passes with TestNA.
See the TEST section in the commit message of CL:1060648 for more details.

Change-Id: I6af1a403981a134c4a6c7a0d3f29456118e89467
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1065060
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>