[autotest] Relocate an update-install sanity check.
During ChromiumOSUpdater.run_update(), after downloading the
new build, but before rebooting, there's a sanity check that
the GPT has been updated to boot the new kernel.  The location
of the check wasn't ideal, since the check is really something
that should be done for every update, not just lab provisioning.

This moves the test to _verify_update_completed(), which is
run after every operation that changes the GPT.

BUG=None
TEST=run provisioning and autoupdate_Rollback in a local instance

Change-Id: I3ceb01e939eee10592768e49f0788c5614074e4e
Reviewed-on: https://chromium-review.googlesource.com/1072868
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>