autotest: Enable parse only mode
This is needed for the Autotest path PROVISIONING -> PARSING (if provision
fails), which was accidentally removed along with the GATHERING code.

This technically relies on a corresponding Lucifer change; otherwise
it will crash due to unrecognized flags.  However, it is already
crashing now since all of the GATHERING code has been removed, so
theres no real point in making sure this lands in order.

BUG=None
TEST=None

Change-Id: Ia71c72d7fa68e8708616dbc73f1e712b31d2c3da
Reviewed-on: https://chromium-review.googlesource.com/1072634
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>