autotest: utils: add function to sample in the bg
There is a need to sample the result of a value in the background of
another running task, such as monitoring bitrates or other connection
information during testing. Add this here so it is accessible. This was
done with chromium:845762 and chromium:845755 in mind but should be more
broadly applicable.

CQ-DEPEND=CL:1071091
BUG=chromium:842219
TEST=python client/common_lib/utils_unittest.py test_background_sample

Change-Id: Ie88551b299e08155999313aff8adb31d113da85d
Reviewed-on: https://chromium-review.googlesource.com/1072739
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>