Add fps_meter to autotest library
This patch is a part of a series of patches. The goal is to move
the fps_meter to autotest library and update the ebuild file so
that it is added to the /usr/local/bin and it can be used as a
tool accordingly. In this particular patch, fps_meter is added to
the autotest library.

BUG=chromium:845378
TEST=Manual tested

Change-Id: Id2d96e6997193a2db035407744aa6a91608f77c8
Reviewed-on: https://chromium-review.googlesource.com/1072971
Commit-Ready: Yi Xu <yiyix@chromium.org>
Tested-by: Yi Xu <yiyix@chromium.org>
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Reviewed-by: Yi Xu <yiyix@chromium.org>