chrome_cr50: prepend '\n' to cr50 commands
We saw that sometimes junk would be sent with real commands. Adding \n
to the front of the command will separate junk that has been sent to the
console from the command we are running. We were seeing ccdstate become
Iccdstate, so this would send I and ccdstate as separate commands.

BUG=b:80201548
BRANCH=none
TEST=run firmware_Cr50CCDServoCap

Change-Id: I6151df769b9aa5318bb672a677518f7259164524
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1074331
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>