network_WiFi_ChaosConfigSniffer: Replace mutable argument list with None
This fixes:
************* Module network_WiFi_ChaosConfigSniffer
W: 19, 4: Dangerous default value [] as argument (dangerous-default-value)

From the python docs:
(https://docs.python.org/3/tutorial/controlflow.html#default-argument-values)
The default value is evaluated only once. This makes a difference when the
default is a mutable object such as a list, dictionary, or instances of most
classes. A function with a mutable argument accumulates the arguments passed
to it on subsequent calls.

BUG=chromium:847272
TEST=None, this test doesn't run on my workstation even without the change.

Change-Id: I0847fe3ec46b4fb19dda0051b9846263e83d9361
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1075781
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>