firmware_ECBootTime: Do not rely on "Console is enabled" string from EC
Current tests requires "Console is enabled" to be printed from EC.
However, depending on how busy the EC is on boot, the string may
be corrupted/interleaved with other strings.

Instead, after matching the first set of "ec_ready" pattern, we
just wait 1 second, ask for the EC version to check if the EC
console is ready, then press power button.

1 second is a reasonable delay, and EC console must be ready by then,
if not, it is perfectly acceptable to fail this test.

We also refactor get_version function out of firmware_FAFTSetup into
the ChromeEC class.

BUG=chromium:846617
TEST=test_that -b $BOARD $IP firmware_ECBootTime

Change-Id: I0faadbff56c372c6be9bd43a2de94f8f1b9125b9
Reviewed-on: https://chromium-review.googlesource.com/1075832
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>