Add ability to wait for Cryptohome to start up
Adds a timeout parameter to the __init__ method of the Cryptohome daemon
client. If set, the call blocks until the daemon is available or the
timeout is reached.

This will be used for enterprise_AuthPolicyDaemon to ensure that
Cryptohome is up. On some boards the test currently fails with
  DBusException: org.freedesktop.DBus.Error.ServiceUnknown:
  The name org.chromium.Cryptohome was not provided by any .service files

CQ-DEPEND=CL:*632267

BUG=chromium:843093
TEST=cros tryjob -g *632267 -g 1076230 eve-paladin-tryjob

Change-Id: Iae1d522893067d0aed51af6e0d7ec32274d3c322
Reviewed-on: https://chromium-review.googlesource.com/1076230
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>