security_RootCA: Format baseline for easier updates
Having a consistent format makes it easier to see the diffs of
programmatic updates. This CL adds a python script that can be run to
re-format an updated baseline, re-formats the current baseline with it,
adds another python script that can be used to update the baseline with
new openssl roots (can be useful at the time of updating
app-misc/ca-certificates), and removes the old bash script seemed to
exist for the same purpose, but was hard to read and extend.

BUG=b:78572490
TEST=The test still passes with the reformatted baseline

Change-Id: I01ab2a99354f340e6f1017ad2fe95f95e9a53e86
Reviewed-on: https://chromium-review.googlesource.com/1083225
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>