security_SandboxedServices: reflect shill sandboxing changes
This CL should land along with enabling spawning of shill in a minijail
as a non-root user with -n (no_new_privs). This change and the dependent
CL are the only 2 things that need to be reverted to go back to running
the shill process tree as root.

CQ-DEPEND=CL:1087540
BUG=chromium:649417
TEST=tests passes alongside larger sandbox shill debug CL

Change-Id: I7a1145ff8ad9f0a2d39d1c432220dc628f3d39e5
Reviewed-on: https://chromium-review.googlesource.com/1082958
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>