FirmwareStatusVerifier: don't create unused file
The previous invocation would create a ROM-sized file (named
image.bin) that was not used.  If the ROM-sized file that flashrom is
asked to create exists (which for the second iteration here, it
would), flashrom looks to look for an FMAP in it when the -i flag is
also used.  That could be a flashrom bug (it would make more sense to
prefer the FMAP on the ROM), but not explored further here, as that
ROM-sized file is not needed for this use case.

TEST=ran "for section in VBLOCK_A VBLOCK_B FW_MAIN_A FW_MAIN_B; do
flashrom -r -i $section:$section;  done;" on an eve and
inspected the output for error messages.
BUG=b:80244293

Change-Id: Ifa2ece8dd1d956564546d723220de783812cec70
Reviewed-on: https://chromium-review.googlesource.com/1087683
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>