network_WiFi_BSSTMReq: Fix test failures
Many of the network_WiFi_BSSTMReq test failures are caused be either (a) the AP not
supporting BSS Transition Management, or (b) the driver picking the original AP to
roam to instead of the one suggested by the BSS Transition Management Request (this
is possibly because the driver prefers the 5GHz band).

To fix (a), I checked for the output of the request ('OK' indicates that it is
supported and was successfully sent) and raised a TestNAError if 'OK' was not found.

To fix (b), I switched the two APs so that the client would instead be roaming from
2.4GHz band to 5GHz band.

BUG=chromium:844762
TEST=Ran on machines that failed these tests (e.g. chromeos2-row11-rack4-host2 and
     chromeos2-row11-rack7-host6) and they now pass.

Change-Id: I1efbd73b8f8285c5c0a7705414d7e6a66747669b
Reviewed-on: https://chromium-review.googlesource.com/1087806
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>